Split login components into separate files
