+++ plugins/authentication/src/ExternalTokenModel/model.tsx
@@ -4,13 +4,8 @@ import { RemoteFileWithRangeCache } from '@jbrowse/core/util/io'
 import { UriLocation } from '@jbrowse/core/util/types'
 import { ExternalTokenInternetAccountConfigModel } from './configSchema'
 import { Instance, types, getParent } from 'mobx-state-tree'
-import React, { useState } from 'react'
-import Button from '@material-ui/core/Button'
-import Dialog from '@material-ui/core/Dialog'
-import DialogContent from '@material-ui/core/DialogContent'
-import DialogTitle from '@material-ui/core/DialogTitle'
-import DialogActions from '@material-ui/core/DialogActions'
-import TextField from '@material-ui/core/TextField'
+
+import { ExternalTokenEntryForm } from './ExternalTokenEntryForm'
 
 const inWebWorker = typeof sessionStorage === 'undefined'
 
@@ -192,60 +187,6 @@ const stateModelFactory = (
         },
       }
     })
-}
-
-const ExternalTokenEntryForm = ({
-  internetAccountId,
-  handleClose,
-}: {
-  internetAccountId: string
-  handleClose: (arg?: string) => void
-}) => {
-  const [token, setToken] = useState('')
-
-  return (
-    <>
-      <Dialog open maxWidth="xl" data-testid="externalToken-form">
-        <DialogTitle>Enter Token for {internetAccountId}</DialogTitle>
-        <DialogContent style={{ display: 'flex', flexDirection: 'column' }}>
-          <TextField
-            required
-            label="Enter Token"
-            variant="outlined"
-            inputProps={{ 'data-testid': 'entry-externalToken' }}
-            onChange={event => {
-              setToken(event.target.value)
-            }}
-            margin="dense"
-          />
-        </DialogContent>
-        <DialogActions>
-          <Button
-            variant="contained"
-            color="primary"
-            type="submit"
-            disabled={!token}
-            onClick={() => {
-              if (token) {
-                handleClose(token)
-              }
-            }}
-          >
-            Add
-          </Button>
-          <Button
-            variant="contained"
-            color="primary"
-            onClick={() => {
-              handleClose()
-            }}
-          >
-            Cancel
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </>
-  )
 }
 
 export default stateModelFactory
